requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
livereload==2.7.1
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
qrcode==8.2
tornado==6.5.4
Werkzeug==3.1.4
